requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
databases==0.9.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.114.0
greenlet==3.0.3
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.9.0
pydantic_core==2.23.2
python-decouple==3.8
python-jose==3.3.0
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
